Reorganize: move VTK code from mesh.cpp to vtk.cpp
