+++ examples/ex5.cpp
@@ -269,7 +269,7 @@ int main(int argc, char *argv[])
    paraview_dc.SetPrefixPath("ParaView");
    paraview_dc.SetLevelsOfDetail(order);
    paraview_dc.SetCycle(0);
-   paraview_dc.SetDataFormat(VTUFormat::BINARY);
+   paraview_dc.SetDataFormat(VTKFormat::BINARY);
    paraview_dc.SetHighOrderOutput(true);
    paraview_dc.SetTime(0.0); // set the time
    paraview_dc.RegisterField("velocity",&u);
